pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "malaysia-prayer-time"
version = "0.2.0"
authors = [
  {name = "Your Name", email = "your-email@example.com"},
]
description = "Malaysia Prayer Time UVX Plugin for Claude Desktop"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "httpx>=0.24.0,<0.25.0",
    "pydantic>=2.0.0,<3.0.0",
    "PyYAML>=6.0,<7.0",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
"Homepage" = "https://github.com/yourusername/mcp-server-malaysia-prayer-time"
"Bug Tracker" = "https://github.com/yourusername/mcp-server-malaysia-prayer-time/issues"

[project.scripts]
mcp-server-waktu-solat = "waktu_solat.server:main"

[project.entry-points."uvx.plugin"]
malaysia-prayer-time = "uvx_plugin"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["waktu_solat"]
include-package-data = true

[tool.setuptools.package-data]
"*" = ["uvx.json"]
